implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -66,7 +66,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-       return view('questions.edit', compact('question'));
+        if(\Gate::denies('update-question', $question)) {
+            abort(403, "Access denied"); 
+        }
+        return view('questions.edit', compact('question'));
     }
 
     /**
@@ -78,6 +81,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question)) {
+            abort(403, "Access denied"); 
+        }
+        return view('questions.edit', compact('question'));   
         $question->update($request->only('title', 'body'));
         return redirect('/questions')->with('success', 'Your question has been updated');
     }
@@ -90,6 +97,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question)) {
+            abort(403, "Access denied"); 
+        }
+        return view('questions.edit', compact('question'));
         $question->delete();
         return redirect('/questions')->with('success', 'Your question has been deleted');
     }
